Generalize section painting
Hunks are still the only sections whose bodies have to painted when
they are highlighted and unhighlighted.  The reason this is necessary
is that highlighting is done by using (an) overlay(s) to change the
background color of selected sections, which shadows an significant
background color used in section bodies.

Hunks use different background colors to indicate added and removed
lines.  Instead of using overlays, the background colors of these
lines (and also of context lines) are changed using text properties,
when a hunk gains or loses focus.  I.e., they are "painted".

Such painting is more expensive than merely removing or adding one
or more overlays, because the text has to be parsed again, so we have
to avoid doing it unless actually necessary.

We do not only want to avoid needlessly painting hunks when moving
around in the buffer, we also want to avoid it for (hunk) sections
whose bodies are not visible yet.  To accomplish that the `washer'
section slot was added.  Its purpose is to delay "washing" until
the section is expanded.

This slot was originally added specifically for hunk sections, but
by now it is used by many other sections too.  In most cases it is
even used to not only delay washing (~= "parsing" and "painting")
but to even delay inserting any section content until that time.

With this commit, we *stop* using the function from the `washer'
slot for hunk painting, and replace that with a new mechanism that
is inspired by the old washer mechanism, but incorporates lessons
learned since.

The two drivers are still `magit-section-update-highlight' and
`magit-section-show'.  These two functions are the main driving
forces behind the two events that make *might* make it necessary
to repaint.

Replace the hook `magit-section-unhighlight-hook' with a new generic
function `magit-section-paint'.  Create such a method for diffs, using
the hunk-specific parts of removed `magit-diff-paint-hunk'.  The parts
that function are not hunk-specific, are replaced with similar logic
in new `magit-section-update-paint'.

`magit-section-update-paint' is where we now decide whether repainting
is actually necessary.