Bring back a variant of delayed hunk refinement
The old variant broke with [1: 95ee9d8c51], which cannot be fixed simply
be reverting that commit, because it was in preparation of subsequent
changes, which heavily refactored section highlighting making it both
simpler and more generic, and also to improve performance.  Delayed hunk
refinement conflicts with the former, and due to the latter it might not
be necessary anymore.

This commit does not fully restore the old behavior.  Previously
delayed refinement meant that only the current hunk was ever refined;
when another section became current, the refinement of the previously
current section was removed again.  Now, once a hunk has been refined,
it remains refined, until the buffer is refreshed.

The reason any form of delayed hunk refinement was implemented in the
first place, was to avoid the upfront cost of completely refining a
large diff.  I never considered this a desirable feature; just a kludge
that was sadly necessary for performance reasons.  IMO the new behavior
is an improvement; though always refining all hunks is of course even
better, unless it is too slow of course.

Closes #5385.

1: 2025-04-26 95ee9d8c5102bc5619f0e2c6d50104cae17b2009
   magit-diff-paint-hunk: Only refine if actually painting