magit-section-update-paint: Register already painted expanded hunk
When we navigate to a collapsed hunk, whose body is secretly already
highlighted and expand it, we can (and already do) avoid painting it
again.  However, we failed to register it as painted again, which is
necessary to cause it to be unhighlighted once we move away from it,
regardless of how far back the painting as highlighted occurred.

We do not just leave the section registered as painted, after collapsing
it and then moving away, because `magit-section-update-highlight' would
then have to consider its state after every command.

Closes #5393.

Suggested-by: Eliza Velasquez