magit-refresh-buffer: Reset magit-section-focused-sections
This variable is also reset in `magit-section-pre-command-hook' but that
only affects the current buffer.  We also have to reset this cache when
refreshing a buffer, otherwise it would continue to track sections from
the previous generation, which no longer exist.

For example, if point is on an unstage hunk in the status and the user
invokes a commit command, then the commit message buffer is current when
the commit is finalized.  In that case `magit-section-update-paint' was
called for two section in the status buffer; the previous generation
hunk (which ended up being a noop) and the actually focused hunk (which
was not included in the value returned by `magit-focused-sections',
because that cache had not been invalidated yet, and thus got painted
as unfocused).