magit-discard: On "Untracked files" delete exactly the listed files
Use the same code to generate a list of untracked fils, when
`magit-discard' is invoked on the "Untracked files" section as
was used to populate the list of files displayed in that section.

We should do that even if it weren't for the following issue, but
that is how we noticed something had to be done.

Contrary to what the documentation claims, "git ls-files --other
--directory --exclude-standard" does not honor ".gitignore" from
sub-directories.

Closes #5405.