rebase: Deal with comment character in the middle of action lines
In "merge" lines the description was always prefixed with "# ".
Since Git v2.50.0 that is also done for other action lines.

To support older releases, "# " has to be optional for these other
action lines.  While that isn't necessary, make it optional for
"merge" lines as well.

Note that Git always uses "#"; `core.commentChar' does not control
the character used here.

Never-the-less start using `font-lock-comment-face' for this inline
"#"; frankly it is unnecessary noise that is best diminished.