git-rebase-mode-font-lock-keywords: Use magit-branch-local face
Use this face for branch names in "# Branch" and "merge" lines, but
do not use it in "label" lines.  The former two mark the points were
a branch is created and merged, while the latter is an implementation,
needed for technical reasons.

Relocate the keyword for "# Branch", so that it is placed among the
action line keywords instead of after the keywords for the comments
that optionally follow the action lines.