Don't kill blob buffers when displayed in multiple windows
Previously "q" in a blob buffer unconditionally killed it.  I have
a personal command that displays the current buffer in a new frame
(while still displaying it in the old window/frame), which I often
follow up by quitting the current buffer in the old window, which
then ends up killing the buffer "in both frames".

On the other hand, when time traveling, it is useful if "q" cleans
up completely, else we would end up with many blob buffers, which
are no longer of interest.

Determining whether to kill or only bury the current buffer based on
whether it is displayed in a second window, covers both use-cases.

Other users might find different behaviors useful, so implement and
mention some alternatives.