magit-diff--locate-file-position: New function
Compared to its predecessor, `magit-diff-locate-hunk', this not only
tries to locate the appropriate hunk, it also takes care of the next
step, which is to locate the position within the hunk, provided we
found the appropriate hunk.  Previously callers had to take care of
that.

Because callers may act differently depending on whether they found
the exact position, the hunk, or just the file, include the exact
locality in the return value.  The predecessor already did this,
except it only used a boolean, indicating whether a hunk was found
or not.  Return the position, not the section.