magit-status-goto-file-position: Favor unstaged hunk
If a matching unstaged hunk is found and the line is modified in that,
use that position.  Otherwise a matching unstaged hunk is found and
the line is modified in that, use that position.  I.e., favor unstaged
changes, but if the line is modified in the staged changes but not in
the unstaged changes, then favor the staged hunk which last modified
the line in question -- not the unstaged hunk that happens to modify
some close by line.

If the line is modified in neither the unstaged nor staged changes,
again favor the former over the latter, but this time ignoring line
precision.

Finally as before, if no hunk containing the line is found, find the
file, again favoring unstaged over staged changes.  As before go either
to the heading of the file or the heading of the last hunk, depending
on whether the line in question comes before or after the modified
lines.

Before this commit we favored staged changes over unstaged changes,
but that made little sense.

Closes #4814.