magit-diff-visit-file--setup: Do not require visible buffer
Do not require a visible window to display the buffer, or any
window for that matter.  Either is unlikely to happen, but given
an odd display function it is theoretically possible.  We can set
point even if no window displays the buffer, so do not needlessly
signal an error should either unexpectedly happen.

[1: 0679c015a9] was right to ensure the correct window is selected
but failing if no visible window displays it went too far.

1: 2017-05-22 0679c015a9255c848cd4d9d8a71828ba45887590
   magit-diff-visit-file: adjust point in correct window