magit-diff-visit-file--internal: Partially respect DISPLAY for directory
Previously we passed along the wrong variable.  If DISPLAY is non-nil,
`magit-diff-visit-directory' displays in another window.  When it is
`switch-to-buffer-other-frame', that isn't quite what the user asked
for but it is better to even use the same window and given that
`magit-diff-visit-directory' does not currently support using another
frame, that is the best we can do.