magit-tag-create: No longer use magit-read-tag
This allows dropping the REQUIRE-MATCH argument from `magit-read-tag',
because now all its callers do require a match (though some failed to
specify so).

Keep using completing read when creating a new tag and offer the
existing tags as "choices".  The idea is that makes it easier to
see what tags already exist and pick another name for the new one.