Use _ in all non-binding entries in the varlist of COND-let forms
This was not done until now because doing so resulted in a warning
because the macro expansion did not actually leave any variables
unused.  This was fixed in Emacs 30.1 and I've added a backport to
Transient, which this package can also take advantage of.

Without consistently using `_' one would always have to count parens
to be sure whether a member of the varlist does or does not bind a
variable.  It is very easy to make mistake when writing or reading
such forms, if `_' is not consistently used in the non-binding cases.