magit-section-mode: Set font-lock-defaults to disable syntactic font-lock
Setting `font-lock-keywords-only' as we did since [1: 9e6791796f]
(and fixed in [2: 649bc9dc32]) does not work because once some
third-party minor-mode calls `font-lock-add-keywords', its value
is set anew (from `font-lock-defaults', which we did not set).

As stated in the manual, such variables should not be set directly,
one should instead set `font-lock-defaults'.  Failure to follow that
advice results in issues such as the above.

We avoided setting `font-lock-defaults' ever since [3: 5a2786cc00],
but that was misguided.  By default `font-look-mode' is enabled for
all major-modes anyway, so this attempt to not enable font-lock in
`magit-section-mode' buffers if at all possible was futile.  And
if it had succeeded that would actually have disabled most if not
all fontification in Magit buffers (because we use `font-lock-face',
(not `face') as we are supposed to).

(A more descriptive name for the `font-lock-face' property would be
`none-font-lock-face-which-font-lock-mode-does-not-discard'.)

Closes #5420.

1: 2025-07-24 9e6791796facd87d6d46abca2628e802edf01ec4
   Set font-lock-keywords-only to disable syntactic fontification

2: 2025-08-13 649bc9dc328b8cb7519292be4f5c9cac13cf3466
   magit-section-mode: Set font-lock-keywords-only locally

3: 2019-08-22 5a2786cc005c1dfb698b6c2fc1dd592c344b0e4a
   magit-mode: Disable syntactic font-lock differently