Revert "Remove kludge for and-let* bug in older Emacs releases"
This reverts commit 92020002b3f162e26af11f1a01c21bb18111b1cc.

These `progn' used to be necessary because `and-let*' used to behave
like `if'.  That bug was fixed in Emacs 27.1 -- since then it behaves
like `when', as its author intended.

However, going forward we will use the `and-let*' implementation from
the `cond-let' package and that takes a single BODYFORM.  Sometimes
it would be convenient if it supported multiple BODY forms and they
behaved like those were wrapped with `progn' and other times like they
were wrapped with `and', and we could determine which would be useful
more often and the optimize for that.  Requiring an explicit `and' or
`progn' is better, especially when the author is forced to call out
the fact that there *also* is a side-effect.