Use Cond-Let's implementations of {if,when,and,while}-let{,*}
Cond-Let's `if-let', `if-let*', `when-let', `when-let*' and `and-let*'
implementations behave differently from the built-in implementations,
but because I have been careful not to depend on the idiosyncrasies,
switching over should be mostly safe.

Also add shorthands for the other `cond-let--...' macros.  We will
start using those (and `cond-let' and `cond-let*') in later commits.