Use cond-let and cond-let*
I've only recently implemented these forms, but before that I have
been hoping they would be added to Emacs for a decade or so.  Even
so, I do not yet have any practice using them, and so my sense of
when to use them, and when not, has yet to evolve.