Use Cond-Let's and$
Anaphoric macros get a bad rep and while my own infatuation was somewhat
short lived, there are cases where I still find them useful, such as in
the case of

  (and$ (function-with-a-name-that-makes-if-obvious-what-is-returned)
        (function-with-a-name-that-makes-if-obvious-what-is-expected $))

Many of the cases where this commits starts using `and$', fall squarely
into that category.  There are a few where I an on the fence.  That's
because I've just invented this macro and need some practice to fully
develop my sense of when it is, and is not, appropriate to use this
macro.