magit-set-window-margins: Fix recent regression
In [1: a47b5098c8] we started to fall back to preserving the current
width for the right margin, similar to how we do for the left column.
That makes sense for the left column, where we want to continue to
support something else making use of that.  However in the case the
right margin Magit claims absolute ownership in Magit buffers.  The
result of this was that we lost the ability to hide the right margin
again.

1: 2025-08-26 a47b5098c812b7a7932df802378cb754e3ad8639
   Support displaying visibility indicator in left margin