magit-diff-visit--position: Don't add bogus offset when visiting index
We only need to calculate the offset when visiting the worktree
because GOTO-FILE is non-nil, despite REV not being "{worktree}".

But we always added an offset when REV is "{index}".  We must only
add an offset when visiting "{worktree}" instead of REV, including
when REV is "{index}".  The only thing special about "{index}" is
that `magit-diff-visit--offset' doesn't understand it, so we have
to substitute nil, when calling that function.