magit-revert-rev-file-buffer: Respect nil enable-local-variables
If the user has set `enable-local-variables' to nil, we want to
respect that.  `normal-mode's FIND-FILE argument wasn't designed for
our use case, so we have to use this strange invocation to achieve
that.

Unfortunately, it is not possible to prevent every non-nil value being
interpreted as t.  In the future we might re-implement `normal-mode'
to respect those values as well.