magit-jump-to-revision-diffstat: Jump to diffstat file from hunk
Before the menu `magit-revision-jump' was added, when point is on a
file in a diff "j" (`magit-jump-to-diffstat-or-diff') jumped to the
respective file within the diffstat.  For `magit-diff-mode' it still
does that.

In `magit-revision-mode' "j" is now bound to `magit-revision-jump',
with partially shadows the above behavior.  To continue to allow
quickly jumping between a file in a revision diff and the same file
listed in the diffstat, `magit-revision-jump' pivots to calling
`magit-jump-to-diffstat-or-diff' when invoked in those locations.

The hope was that would preserve enough of the old behavior to make
everyone happy.  I.e., on a file jump to the other place that file
is listed (as before), everywhere else (or with a prefix argument)
show the menu, letting the user explicitly choose where to jump to
this time around (previously this would have jumped to the diffstat).

Users who want the old behavior back can do:

  (keymap-unset magit-revision-mode-map "j" t)

Instead of adding a more explicit knob to restore the old behavior,
this commit improves a variation:

If `magit-jump-revision-diffstat' is invoked on a hunk (via "j s"),
that now goes to the respective file in the diffstat, instead of to
the heading of the diffstat itself.

The hope is that going to a more precise location outweighs the added
cost of having to press "j s" instead of just "j" to go to the diffstat.

Do the same for `magit-jump-to-revision-diffstat', so that those users,
who remove the new binding in `magit-revision-mode-map', can benefit
too.

Now let's just hope nobody dislikes going to the file heading instead of
to the diffstat heading.

Closes #5442.