magit-completing-read-multiple: Restore meaning of t as REQUIRE-MATCH
As for `completing-read-multiple', if REQUIRE-MATCH is t, then all
inputs must appear in the collection, but it is again possible to
select nothing.

Forge's uses of this function need this behavior.  In [1: 4e2dd7ba6f]
I should only have added the new `any' value, instead of also changing
the meaning of t.

Closes magit/forge#814.

1: 2025-08-16 4e2dd7ba6f54b2251a702b4c0416467d7315407a
   magit-completing-read-multiple: Support requiring non-empty input