magit-diff-wash-diff: Also stop when reaching a log heading
A diff may contain just headers and no hunk.  When washing a diff
directly, we already detected that by checking whether we have reached
the beginning of the next diff or eob.

When the diff is part of a log, we additionally have to also check
whether we have reached the next log heading.

Alternatively we could pass along the limit from `magit-log-wash-rev'
(as a marker, because we modify the buffer text), but that would
complicate matters.

Closes #5454.