Support Emacs builds that feature custom-initialize-after-file-load
This will need more work.  We might be able to learn from that function
and simplify even when that is not available.  Also, while rudimentary
testing indicates this works as intended, this has to be studied in more
detail before we can be sure.  Do it now anyway, because the breakage
caused by not doing is definitely real.

See #5462.