magit-section-visibility-indicator: Optimize display of custom values
For constant choices, we would like to display just the "tag" because
displaying the value would be redundant.  For composed types we cannot
just drop "%v" from the format.  The child values have to represent
by some widget/text in the buffer, because the widget library expects
that, even if the values of the child widgets are constant.  So insert
" " for each cons cell slot.