magit-auto-revert-mode: Improve initialization
Use new initialization function `magit-custom-initialize-after-init',
which was inspired by the new `custom-initialize-after-file-load' in
Emacs 31.1.

It is necessary to continue to delay mode initialization until after
Emacs initialization has completed, because enabling this mode causes
`with-editor' to be required, and if Emacs initialization isn't done,
then `with-editor-autoloads' most likely hasn't been loaded yet
(because autoload files are loaded in alphabetic order), and that
likely means that `with-editor' is not on the `load-path' yet.