Drop a space from autoload cookies
The space (if any) between the keyword and form is not stripped
when inserting the form into the autoloads file, which looks odd.

In Emacs the number of autoload cookies with and without such a
space, is almost the same, so that provides no guidance as to what
should be preferred.  Go with what makes the aligner in me happy.