magit-insert-worktrees: mention in documentation
This section inserter is handy, and has existed since 2.7.0 /
82bfdd46295c5d6fa5d1a26a5058994e5b4f7c56... but appears to not yet be
documented.

This documentation is modeled in the style of
`magit-sparse-checkout-insert-header', where it is mentioned in the
docs for the relevant commands rather than the docs of `magit-status',
since most people who wish see a list of worktrees are likely using
`magit-worktree' / reading its documentation.